reorganize, make close button into a component
